seperate http endpoint

+++ backend/chatservice/dao/db/sqlite/scripts/migrations/3_project_docs_table.up.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS project_docs (
     project_id INTEGER,
     docs_id TEXT,
     file_name TEXT,
-    file_size REAL NOT NULL,
+    file_size INTEGER NOT NULL,
     created_at DATETIME DEFAULT CURRENT_TIMESTAMP,
     updated_at DATETIME DEFAULT CURRENT_TIMESTAMP
 );
